Add clarifying comments defining what a Loop is
Our loop construct is not a way to identify cycles in the CFG.  This wasn't immediately obvious from the header, so clarify that fact.

The motivation for this was that I just fixed a out of tree bug due to a mistaken assumption (on my part) on what a Loop actually was.  While it was fresh in my mind, I wanted to document the key point.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@258154 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit c7630ca0454f6dd51b595cf306f420c3d375ac9c)