[mips] Allowed dla instructions on 32-bit architectures.
(This cherry-pick introduces a helper needed for r263139)

Summary:
This is now the same as the behaviour of the GNU assembler. This was done
as it is required in order to build the Linux kernel with the integrated
assembler enabled.

Reviewers: dsanders, vkalintiris

Subscribers: dsanders, llvm-commits

Differential Revision: http://reviews.llvm.org/D13594

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@258400 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 5e2e572fda827d305ff5327a60ea36df6eb6b1b3)