Revert "[MachO] Stop generating *coal* sections."
This reverts commit c770c8d638a69cd3cd527c54f08008f45bbab2c5.

It turns out that clang attaches section name __datacoal_nt to some of
the global varibles it creates. This causes the variables to be placed
in the datacoal_nt section instead of the data section, which causes
warning "section "__datacoal_nt" is deprecated" to be issued when
an objective-c code is compiled with -save-temps.

rdar://problem/24497535