[X86] Correctly select registers to pop into for x86_64
When trying to replace an add to esp with pops, we need to choose dead
registers to pop into. Registers clobbered by the call and not imp-def'd
by it should be safe. Except that it's not enough to check the register
itself isn't defined, we also need to make sure no overlapping registers
are defined either.

This fixes PR26711.

Differential Revision: http://reviews.llvm.org/D18029

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@263139 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit b448651c49a734394ec06b50bba2d12520b44b70)